// Portal SAE/frontend/src/app/interfaces/event/createEvent.dto.ts
export interface CreateEventDTO {
  asunto: string;
  porcentajeProgreso: string;
  fechaHoraInicio: string;
  fechaHoraFin: string;
  tiempoDuracion: number;
  descripcion: string;
  codigoConsultor: string;
  codigoEmpresario: string;
  estado: string;
  tipo: string;
  visibilidad: string;
  notificarPor: string;
  cuentaAsignada: string;
  contactoAsignado: string;
}