// Portal SAE/frontend/src/app/interfaces/event/event.dto.ts
export interface EventDTO {
  codigo: number;
  asunto: string;
  porcentajeProgreso: string;
  fechaHoraInicio: string;
  fechaHoraFin: string;
  tiempoDuracion: number;
  descripcion: string;
  codigoConsultor: string;
  codigoEmpresario: string;
  estado: string;
  tipo: string;
  visibilidad: string;
  notificarPor: string;
  cuentaAsignada: string;
  contactoAsignado: string;
}